후기 사진 파일업로드 기능, 수정/삭제 기능 구현중

+++ src/main/java/com/nadri/attr/vo/Attraction.java
@@ -38,7 +38,7 @@ public class Attraction {
 	private String address;
 	private String detail;
 	private int count;
-	private int star;
+	private double star;
 	
 	private AttOption attOption;
